+++ tests/core/test_playback.py
@@ -50,8 +50,14 @@ class CorePlaybackTest(unittest.TestCase):
         self.unplayable_tl_track = self.tl_tracks[2]
         self.duration_less_tl_track = self.tl_tracks[4]
 
+    def trigger_end_of_track(self):
+        self.core.playback._on_end_of_track()
+
+    def set_current_tl_track(self, tl_track):
+        self.core.playback._set_current_tl_track(tl_track)
+
     def test_get_current_tl_track_none(self):
-        self.core.playback.set_current_tl_track(None)
+        self.set_current_tl_track(None)
 
         self.assertEqual(
             self.core.playback.get_current_tl_track(), None)
@@ -214,7 +220,7 @@ class CorePlaybackTest(unittest.TestCase):
         self.playback2.pause.assert_called_once_with()
 
     def test_pause_changes_state_even_if_track_is_unplayable(self):
-        self.core.playback.current_tl_track = self.unplayable_tl_track
+        self.set_current_tl_track(self.unplayable_tl_track)
         self.core.playback.pause()
 
         self.assertEqual(self.core.playback.state, core.PlaybackState.PAUSED)
@@ -257,7 +263,7 @@ class CorePlaybackTest(unittest.TestCase):
         self.playback2.resume.assert_called_once_with()
 
     def test_resume_does_nothing_if_track_is_unplayable(self):
-        self.core.playback.current_tl_track = self.unplayable_tl_track
+        self.set_current_tl_track(self.unplayable_tl_track)
         self.core.playback.state = core.PlaybackState.PAUSED
         self.core.playback.resume()
 
@@ -300,7 +306,7 @@ class CorePlaybackTest(unittest.TestCase):
         self.playback2.stop.assert_called_once_with()
 
     def test_stop_changes_state_even_if_track_is_unplayable(self):
-        self.core.playback.current_tl_track = self.unplayable_tl_track
+        self.set_current_tl_track(self.unplayable_tl_track)
         self.core.playback.state = core.PlaybackState.PAUSED
         self.core.playback.stop()
 
@@ -419,7 +425,7 @@ class CorePlaybackTest(unittest.TestCase):
         tl_track = self.tl_tracks[0]
         self.core.playback.play(tl_track)
 
-        self.core.playback.on_end_of_track()
+        self.trigger_end_of_track()
 
         self.assertIn(tl_track, self.core.tracklist.tl_tracks)
 
@@ -428,7 +434,7 @@ class CorePlaybackTest(unittest.TestCase):
         self.core.playback.play(tl_track)
         self.core.tracklist.consume = True
 
-        self.core.playback.on_end_of_track()
+        self.trigger_end_of_track()
 
         self.assertNotIn(tl_track, self.core.tracklist.tl_tracks)
 
@@ -438,7 +444,7 @@ class CorePlaybackTest(unittest.TestCase):
         self.core.playback.play(self.tl_tracks[0])
         listener_mock.reset_mock()
 
-        self.core.playback.on_end_of_track()
+        self.trigger_end_of_track()
 
         self.assertListEqual(
             listener_mock.send.mock_calls,
@@ -495,7 +501,7 @@ class CorePlaybackTest(unittest.TestCase):
         self.playback2.seek.assert_called_once_with(10000)
 
     def test_seek_fails_for_unplayable_track(self):
-        self.core.playback.current_tl_track = self.unplayable_tl_track
+        self.set_current_tl_track(self.unplayable_tl_track)
         self.core.playback.state = core.PlaybackState.PLAYING
         success = self.core.playback.seek(1000)
 
@@ -504,7 +510,7 @@ class CorePlaybackTest(unittest.TestCase):
         self.assertFalse(self.playback2.seek.called)
 
     def test_seek_fails_for_track_without_duration(self):
-        self.core.playback.current_tl_track = self.duration_less_tl_track
+        self.set_current_tl_track(self.duration_less_tl_track)
         self.core.playback.state = core.PlaybackState.PLAYING
         success = self.core.playback.seek(1000)
 
@@ -554,7 +560,7 @@ class CorePlaybackTest(unittest.TestCase):
         self.playback2.get_time_position.assert_called_once_with()
 
     def test_time_position_returns_0_if_track_is_unplayable(self):
-        self.core.playback.current_tl_track = self.unplayable_tl_track
+        self.set_current_tl_track(self.unplayable_tl_track)
 
         result = self.core.playback.time_position